requirements.txt
pandas
pyopenssl
scapy
sklearn
tldextract
joblib
